fix: Theme file moved in error
