+++ editor.planx.uk/src/ui/shared/Palette.stories.tsx
@@ -4,7 +4,7 @@ import { useTheme } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
 import { Meta } from "@storybook/react";
 import React from "react";
-import { defaultTheme } from "ui/editor/theme";
+import { defaultTheme } from "theme";
 
 // The `PaletteOptions` exported by Material-UI are expressed as an interface,
 // so this gets us a union type
